+++ middlewares/articleMiddleware.js
@@ -37,4 +37,24 @@ const checkTags = (req, res, next) => {
   return next();
 };
 
-export { checkArticle, checkTags };
+const { tags } = models;
+
+const getTagId = async (req, res, next) => {
+  const tag = await tags.findAndCountAll({
+    where: {
+      tagName: req.query.tag
+    },
+  });
+  if (tag.count < 1) {
+    return StatusResponse.notfound(res, {
+      message: 'No Articles with such tags',
+    });
+  }
+  req.tagId = tag.rows[0].id;
+  req.app.locals.tag = {
+    tagId: req.tagId,
+  };
+  return next();
+};
+
+export { checkArticle, checkTags, getTagId };
